requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Celery
celery==5.3.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# AI & LangChain
langchain==0.1.4
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Web Scraping & Analysis
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.0
httpx==0.26.0
validators==0.22.0
python-whois==0.8.0

# PDF Generation
weasyprint==60.2
reportlab==4.0.9
pypdf==3.17.4

# Google Drive API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Image Processing
Pillow==10.2.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-slugify==8.0.1
python-dateutil==2.8.2
aiofiles==23.2.1

# Rate Limiting
slowapi==0.1.9

# Templating
jinja2==3.1.3

# Monitoring (Optional)
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
